Lunify/requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate==2.7.0
Flask-Script
Flask-Login
psycopg2